move all services to one service dir in product
