+++ admin-dev/themes/new-theme/js/pages/product/combination/combination-list/bulk-delete-handler.ts
@@ -26,7 +26,7 @@
 import {ConfirmModal} from '@components/modal';
 import ProductMap from '@pages/product/product-map';
 import ProductEvents from '@pages/product/product-event-map';
-import CombinationsService from '@pages/product/combination/service/combinations-service';
+import CombinationsService from '@pages/product/service/combinations-service';
 import BulkChoicesSelector from '@pages/product/combination/combination-list/bulk-choices-selector';
 import {EventEmitter} from 'events';
 import ProgressModal from '@components/modal/progress-modal';
