+++ admin-dev/themes/new-theme/js/pages/product/combination/combination-list/bulk-edition-handler.ts
@@ -26,7 +26,7 @@
 import {FormIframeModal} from '@components/modal';
 import ProductMap from '@pages/product/product-map';
 import ProductEvents from '@pages/product/product-event-map';
-import CombinationsService from '@pages/product/combination/service/combinations-service';
+import CombinationsService from '@pages/product/service/combinations-service';
 import {EventEmitter} from 'events';
 import BulkChoicesSelector from '@pages/product/combination/combination-list/bulk-choices-selector';
 import ProgressModal from '@components/modal/progress-modal';
