+++ admin-dev/themes/new-theme/js/pages/product/combination/combination-list/index.ts
@@ -25,16 +25,16 @@
 
 import ProductMap from '@pages/product/product-map';
 import CombinationsListRenderer from '@pages/product/combination/combination-list/combinations-list-renderer';
-import CombinationsService from '@pages/product/combination/service/combinations-service';
+import CombinationsService from '@pages/product/service/combinations-service';
 import DynamicPaginator from '@components/pagination/dynamic-paginator';
 import ProductEventMap from '@pages/product/product-event-map';
 import initCombinationModal from '@pages/product/combination/combination-modal';
 import initFilters from '@pages/product/combination/filters';
 import {EventEmitter} from 'events';
 import initCombinationGenerator from '@pages/product/combination/generator';
-import {getProductAttributeGroups} from '@pages/product/combination/service/attribute-groups';
+import {getProductAttributeGroups} from '@pages/product/service/attribute-groups';
 import BulkEditionHandler from '@pages/product/combination/combination-list/bulk-edition-handler';
-import PaginatedCombinationsService from '@pages/product/combination/service/paginated-combinations-service';
+import PaginatedCombinationsService from '@pages/product/service/paginated-combinations-service';
 import BulkDeleteHandler from '@pages/product/combination/combination-list/bulk-delete-handler';
 import BulkChoicesSelector from '@pages/product/combination/combination-list/bulk-choices-selector';
 import ProductFormModel from '@pages/product/edit/product-form-model';
