+++ admin-dev/themes/new-theme/js/pages/product/edit/manager/attachments-manager.ts
@@ -26,7 +26,7 @@
 import ProductMap from '@pages/product/product-map';
 import ProductEventMap from '@pages/product/product-event-map';
 import Router from '@components/router';
-import {getAttachmentInfo} from '@pages/product/attachment/attachments-service';
+import {getAttachmentInfo} from '@pages/product/service/attachments-service';
 import {FormIframeModal} from '@components/modal';
 import EntitySearchInput from '@components/entity-search-input';
 import {EventEmitter} from 'events';
