+++ admin-dev/themes/new-theme/js/pages/product/service/attachments-service.ts
