+++ admin-dev/themes/new-theme/js/pages/product/service/images-service.ts
