+++ admin-dev/themes/new-theme/js/pages/product/service/paginated-catalog-price-rules-service.ts
